requirements.txt
openai>=1.12.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pillow>=10.0.0
pdf2image>=1.16.0
rich>=13.0.0
reportlab>=4.0.0
fastapi>=0.115.0
uvicorn>=0.34.0
python-multipart>=0.0.20
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.13.0
docling>=2.70.0
ollama>=0.4.0
httpx>=0.27.0
